requirements_fastapi.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.8
jinja2==3.1.2
python-multipart==0.0.6